edubot-api/requirements.txt
fastapi
uvicorn
openai
sqlmodel
sqlalchemy
databases
python-dotenv